Replace PaneUtils with simple function in panes/index.js
Closes #27
